# .specify/specs/wave-0/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SENTINEL — AI-Native Engineering Operations Platform
  description: |
    Unified API for incident management, log monitoring, deployment tracking,
    AI-assisted debugging, and team collaboration.
  version: 1.0.0
  contact:
    name: SENTINEL Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.sentinel.dev/api/v1
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string

    # ──────────────────────────────────────────────
    # Teams & Auth
    # ──────────────────────────────────────────────
...
    Role:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        description:
          type: string
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        team_id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        name:
          type: string
        avatar_url:
          type: string
          nullable: true
        role_id:
          $ref: '#/components/schemas/UUID'
        role:
          $ref: '#/components/schemas/Role'
        is_active:
          type: boolean
        last_login_at:
          $ref: '#/components/schemas/Timestamp'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [email, password, name, team_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        team_name:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    # ──────────────────────────────────────────────
    # Incidents
    # ──────────────────────────────────────────────
    SeverityLevel:
      type: string
      enum: [SEV1, SEV2, SEV3, SEV4]
    IncidentStatus:
      type: string
      enum: [detected, triaging, investigating, mitigating, resolved, closed]
    Incident:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        team_id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        escalated_to:
          $ref: '#/components/schemas/UUID'
        detected_at:
          $ref: '#/components/schemas/Timestamp'
        triaged_at:
          $ref: '#/components/schemas/Timestamp'
        resolved_at:
          $ref: '#/components/schemas/Timestamp'
        closed_at:
          $ref: '#/components/schemas/Timestamp'
        root_cause:
          type: string
        ai_summary:
          type: string
        ai_root_cause_ranking:
          type: array
          items:
            type: object
            properties:
              cause:
                type: string
              confidence:
                type: number
        metadata:
          type: object
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    IncidentCreate:
      type: object
      required: [title, severity]
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        metadata:
          type: object
    IncidentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        escalated_to:
          $ref: '#/components/schemas/UUID'
        root_cause:
          type: string

    # ──────────────────────────────────────────────
    # Logs
    # ──────────────────────────────────────────────
...
    TimelineEvent:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        incident_id:
          $ref: '#/components/schemas/UUID'
        event_type:
          type: string
        source:
          type: string
        actor:
          type: string
        ts:
          $ref: '#/components/schemas/Timestamp'
        payload_ref:
          type: string
        description:
          type: string
        metadata:
          type: object
        created_at:
          $ref: '#/components/schemas/Timestamp'
    TimelineEventCreate:
      type: object
      required: [event_type, source, actor]
      properties:
        event_type:
          type: string
        source:
          type: string
        actor:
          type: string
        payload_ref:
          type: string
        description:
          type: string
        metadata:
          type: object

    # ──────────────────────────────────────────────
    # Tasks
    # ──────────────────────────────────────────────
...
    IncidentMetrics:
      type: object
      properties:
        total_incidents:
          type: integer
        by_severity:
          type: object
          properties:
            SEV1:
              type: integer
            SEV2:
              type: integer
            SEV3:
              type: integer
            SEV4:
              type: integer
        by_status:
          type: object
        mttr_minutes:
          type: number
        sla_breach_rate:
          type: number
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user and team
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [auth]
      summary: Login and get tokens
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /incidents:
    get:
      tags: [incidents]
      summary: List incidents
      operationId: listIncidents
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/SeverityLevel'
        - name: assigned_to
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Incidents list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [incidents]
      summary: Create incident
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incidents/{incident_id}:
    get:
      tags: [incidents]
      summary: Get incident details
      operationId: getIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Incident details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'

    patch:
      tags: [incidents]
      summary: Update incident
      operationId: updateIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentUpdate'
      responses:
        '200':
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incidents/{incident_id}/assign:
    post:
      tags: [incidents]
      summary: Assign incident to user
      operationId: assignIncident
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Incident assigned
...
  /incidents/{incident_id}/timeline:
    get:
      tags: [timeline]
      summary: Get incident timeline
      operationId: getIncidentTimeline
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Timeline events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimelineEvent'

    post:
      tags: [timeline]
      summary: Add timeline event
      operationId: addTimelineEvent
      parameters:
        - name: incident_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimelineEventCreate'
      responses:
        '201':
          description: Event added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEvent'

  # ──────────────────────────────────────────────
  # Tasks
  # ──────────────────────────────────────────────
...
  /analytics/incidents:
    get:
      tags: [analytics]
      summary: Get incident metrics
      operationId: getIncidentMetrics
      parameters:
        - name: start_time
          in: query
          schema:
            $ref: '#/components/schemas/Timestamp'
        - name: end_time
          in: query
          schema:
            $ref: '#/components/schemas/Timestamp'
      responses:
        '200':
          description: Incident metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentMetrics'
...